requirements.txt
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
